admin is asked to confirm before deleting media

+++ src/components/MediaForm.js
@@ -52,30 +52,29 @@ export default function MediaForm(props) {
       props.saveMedia();
     })
     .catch(err => {
-      console.log("ERROR", err)
       setError(err.message);
   })
   };
 
   const handleDeleteMedia = e => {
     e.preventDefault();
-    fetch(`${API_BASE_URL}/media/${props.currentMedia.id}`, {
-      method: 'DELETE',
-      headers: {
-        'Accept': 'application/json',
-        Authorization: `Bearer ${props.authToken}`,
-        'Content-Type': 'application/json'
-      },
+    if(window.confirm('Are you sure you want to delete this media?')){
+      fetch(`${API_BASE_URL}/media/${props.currentMedia.id}`, {
+        method: 'DELETE',
+        headers: {
+          'Accept': 'application/json',
+          Authorization: `Bearer ${props.authToken}`,
+          'Content-Type': 'application/json'
+        },
+      })
+      .then(res => normalizeResponseErrors(res))
+      .then((media) => {
+        props.cancelMedia();
+      })
+      .catch(err => {
+        setError(err.message);
     })
-    .then(res => normalizeResponseErrors(res))
-    // .then(res => res.json())
-    .then((media) => {
-      props.cancelMedia();
-    })
-    .catch(err => {
-      console.log("ERROR", err)
-      setError(err.message);
-  })
+    }
   };
   
   return (
